refactor: modularize converters module into separate components
Split the monolithic converters.rs (1635 lines) into focused modules:
- patterns.rs: Pattern variable extraction for match expressions
- variables.rs: Variable index adjustment for closure environments
- quasiquote.rs: Quasiquote expansion and handling
- value_to_expr_internal.rs: Main value-to-expression conversion (1200+ lines)
  Handles all special forms: lambda, let/let*/letrec, if, cond, try/catch,
  match, quotes, macros, threading operators (-> and ->>), and more
- tail_calls.rs: Tail call position marking for TCO optimization
- mod.rs: Public API and module exports

This is the largest refactoring to date, breaking down complex compilation
logic into digestible, focused modules.

Benefits:
- Clear separation of concerns across compilation stages
- value_to_expr_internal can be further split if needed
- Each module has a single, well-defined responsibility
- Easier to understand the overall compilation flow
- Better code navigation and maintainability

Testing:
✅ All 1395 tests pass
✅ No functional changes
✅ Clippy: clean
✅ Format: verified
