+++ src/compiler/converters/value_to_expr_internal.rs
@@ -1,336 +1,16 @@
-use super::analysis::{analyze_capture_usage, analyze_free_vars};
-use super::ast::{Expr, Pattern};
-use super::macros::expand_macro;
-use super::patterns::value_to_pattern;
+use super::super::analysis::{analyze_capture_usage, analyze_free_vars};
+use super::super::ast::Expr;
+use super::super::macros::expand_macro;
+use super::patterns::extract_pattern_variables;
+use super::quasiquote::expand_quasiquote;
+use super::variables::adjust_var_indices;
 use crate::symbol::SymbolTable;
 use crate::value::{SymbolId, Value};
 use std::collections::HashSet;
 
-/// Extract all variable bindings from a pattern
-fn extract_pattern_variables(pattern: &Pattern) -> Vec<SymbolId> {
-    let mut vars = Vec::new();
-    match pattern {
-        Pattern::Var(sym_id) => {
-            vars.push(*sym_id);
-        }
-        Pattern::Wildcard | Pattern::Literal(_) | Pattern::Nil => {
-            // These don't bind variables
-        }
-        Pattern::List(patterns) => {
-            // Extract variables from all list elements
-            for p in patterns {
-                vars.extend(extract_pattern_variables(p));
-            }
-        }
-        Pattern::Cons { head, tail } => {
-            // Extract from head and tail
-            vars.extend(extract_pattern_variables(head));
-            vars.extend(extract_pattern_variables(tail));
-        }
-        Pattern::Guard { pattern: inner, .. } => {
-            // Extract from inner pattern
-            vars.extend(extract_pattern_variables(inner));
-        }
-    }
-    vars
-}
-
-/// Adjust variable indices in an expression to account for the closure environment layout.
-/// The closure environment is laid out as [captures..., parameters...]
-///
-/// During parsing, Var nodes contain indices relative to the scope_stack at parse time.
-/// After computing captures, we need to convert these to absolute indices in the final environment.
-///
-/// For each Var:
-/// - If it's a parameter of the current lambda: map to captures.len() + position_in_params
-/// - If it's a captured value: map to position_in_captures_list
-fn adjust_var_indices(expr: &mut Expr, captures: &[(SymbolId, usize, usize)], params: &[SymbolId]) {
-    // Build a map of captures to their position in the list
-    let mut capture_map: std::collections::HashMap<SymbolId, usize> =
-        std::collections::HashMap::new();
-    for (i, (sym, _, _)) in captures.iter().enumerate() {
-        capture_map.insert(*sym, i);
-    }
-
-    // Build a map of parameters to their position in the list
-    let mut param_map: std::collections::HashMap<SymbolId, usize> =
-        std::collections::HashMap::new();
-    for (i, sym) in params.iter().enumerate() {
-        param_map.insert(*sym, i);
-    }
-
-    match expr {
-        Expr::Var(sym_id, _depth, index) => {
-            // Adjust indices for variables that are captures or parameters of this lambda
-            // Both depth==0 (current scope params) and depth>0 (captured vars) need adjustment
-            // because the index is currently relative to scope_stack at parse time, not the
-            // final closure environment [captures..., parameters...]
-            if let Some(cap_pos) = capture_map.get(sym_id) {
-                // This variable is a capture - map to its position in the captures list
-                *index = *cap_pos;
-            } else if let Some(param_pos) = param_map.get(sym_id) {
-                // This variable is a parameter - map to captures.len() + position
-                *index = captures.len() + param_pos;
-            }
-            // Otherwise, it's a global or something from an even outer scope
-        }
-        Expr::If { cond, then, else_ } => {
-            adjust_var_indices(cond, captures, params);
-            adjust_var_indices(then, captures, params);
-            adjust_var_indices(else_, captures, params);
-        }
-        Expr::Cond { clauses, else_body } => {
-            for (test, body) in clauses {
-                adjust_var_indices(test, captures, params);
-                adjust_var_indices(body, captures, params);
-            }
-            if let Some(else_expr) = else_body {
-                adjust_var_indices(else_expr, captures, params);
-            }
-        }
-        Expr::Begin(exprs) => {
-            for e in exprs {
-                adjust_var_indices(e, captures, params);
-            }
-        }
-        Expr::Block(exprs) => {
-            for e in exprs {
-                adjust_var_indices(e, captures, params);
-            }
-        }
-        Expr::Call { func, args, .. } => {
-            adjust_var_indices(func, captures, params);
-            for arg in args {
-                adjust_var_indices(arg, captures, params);
-            }
-        }
-        Expr::Lambda { .. } => {
-            // Don't adjust nested lambda bodies at all - they have already been fully processed
-            // with their own capture scopes, parameters, and indices when they were parsed.
-            // Recursing into them would incorrectly adjust already-correct indices.
-        }
-        Expr::Let { bindings, body } => {
-            for (_, expr) in bindings {
-                adjust_var_indices(expr, captures, params);
-            }
-            adjust_var_indices(body, captures, params);
-        }
-        Expr::Letrec { bindings, body } => {
-            for (_, expr) in bindings {
-                adjust_var_indices(expr, captures, params);
-            }
-            adjust_var_indices(body, captures, params);
-        }
-        Expr::Set { value, .. } => {
-            adjust_var_indices(value, captures, params);
-        }
-        Expr::While { cond, body } => {
-            adjust_var_indices(cond, captures, params);
-            adjust_var_indices(body, captures, params);
-        }
-        Expr::For { iter, body, .. } => {
-            adjust_var_indices(iter, captures, params);
-            adjust_var_indices(body, captures, params);
-        }
-        Expr::Match {
-            value,
-            patterns,
-            default,
-        } => {
-            adjust_var_indices(value, captures, params);
-            for (_, expr) in patterns {
-                adjust_var_indices(expr, captures, params);
-            }
-            if let Some(default_expr) = default {
-                adjust_var_indices(default_expr, captures, params);
-            }
-        }
-        Expr::Try {
-            body,
-            catch,
-            finally,
-        } => {
-            adjust_var_indices(body, captures, params);
-            if let Some((_, handler)) = catch {
-                adjust_var_indices(handler, captures, params);
-            }
-            if let Some(finally_expr) = finally {
-                adjust_var_indices(finally_expr, captures, params);
-            }
-        }
-        Expr::And(exprs) | Expr::Or(exprs) => {
-            for e in exprs {
-                adjust_var_indices(e, captures, params);
-            }
-        }
-        Expr::DefMacro { body, .. } => {
-            adjust_var_indices(body, captures, params);
-        }
-        // Literals, GlobalVar, etc. don't need adjustment
-        _ => {}
-    }
-}
-
-/// Expand a quasiquote form with support for unquote and unquote-splicing
-/// Quasiquote recursively quotes all forms except those wrapped in unquote/unquote-splicing
-fn expand_quasiquote(
-    value: &Value,
-    symbols: &mut SymbolTable,
-    scope_stack: &mut Vec<Vec<SymbolId>>,
-) -> Result<Expr, String> {
-    // For now, implement quasiquote as a simple transformation:
-    // `(a ,x ,@y b) becomes (list 'a x y 'b)
-    // We'll build an expression that constructs the list at runtime
-
-    build_quasiquote_expr(value, symbols, scope_stack, 1)
-}
-
-/// Build an expression that will construct the quasiquoted list at runtime
-fn build_quasiquote_expr(
-    value: &Value,
-    symbols: &mut SymbolTable,
-    scope_stack: &mut Vec<Vec<SymbolId>>,
-    depth: usize,
-) -> Result<Expr, String> {
-    match value {
-        // Simple values are quoted
-        Value::Nil
-        | Value::Bool(_)
-        | Value::Int(_)
-        | Value::Float(_)
-        | Value::String(_)
-        | Value::Symbol(_)
-        | Value::Keyword(_) => Ok(Expr::Literal(value.clone())),
-
-        Value::Table(_) | Value::Struct(_) => Ok(Expr::Literal(value.clone())),
-
-        Value::Cons(_) => {
-            let list = value.list_to_vec()?;
-            if list.is_empty() {
-                return Ok(Expr::Literal(Value::Nil));
-            }
-
-            // Check for special forms
-            if let Value::Symbol(sym) = &list[0] {
-                let name = symbols.name(*sym).ok_or("Unknown symbol")?;
-
-                match name {
-                    // Nested quasiquote
-                    "quasiquote" => {
-                        if list.len() != 2 {
-                            return Err("quasiquote requires exactly 1 argument".to_string());
-                        }
-                        build_quasiquote_expr(&list[1], symbols, scope_stack, depth + 1)
-                    }
-
-                    // Unquote - evaluate the expression
-                    "unquote" => {
-                        if list.len() != 2 {
-                            return Err("unquote requires exactly 1 argument".to_string());
-                        }
-                        if depth == 1 {
-                            // Evaluate this expression
-                            value_to_expr_with_scope(&list[1], symbols, scope_stack)
-                        } else {
-                            // Nested unquote - decrease depth
-                            build_quasiquote_expr(&list[1], symbols, scope_stack, depth - 1)
-                        }
-                    }
-
-                    // Unquote-splicing only valid in list context
-                    "unquote-splicing" => Err(
-                        "unquote-splicing can only be used inside a quasiquoted list".to_string(),
-                    ),
-
-                    // Regular list - recursively process
-                    _ => {
-                        // Process elements and build a list construction
-                        let mut elements = Vec::new();
-                        for elem in &list {
-                            // Check if this is unquote-splicing
-                            if let Value::Cons(_) = elem {
-                                if let Ok(elem_vec) = elem.list_to_vec() {
-                                    if !elem_vec.is_empty() {
-                                        if let Value::Symbol(elem_sym) = &elem_vec[0] {
-                                            if let Some(elem_name) = symbols.name(*elem_sym) {
-                                                if elem_name == "unquote-splicing" && depth == 1 {
-                                                    // Mark for splicing
-                                                    if elem_vec.len() != 2 {
-                                                        return Err(
-                                                            "unquote-splicing requires 1 argument"
-                                                                .to_string(),
-                                                        );
-                                                    }
-                                                    elements.push(Expr::Literal(Value::Symbol(
-                                                        symbols.intern("__splice__"),
-                                                    )));
-                                                    elements.push(value_to_expr_with_scope(
-                                                        &elem_vec[1],
-                                                        symbols,
-                                                        scope_stack,
-                                                    )?);
-                                                    continue;
-                                                }
-                                            }
-                                        }
-                                    }
-                                }
-                            }
-
-                            // Regular element
-                            elements.push(build_quasiquote_expr(
-                                elem,
-                                symbols,
-                                scope_stack,
-                                depth,
-                            )?);
-                        }
-
-                        // For now, just return the list of elements as a literal
-                        // A full implementation would need runtime support for splicing
-                        let mut result_list = Vec::new();
-                        for elem_expr in elements {
-                            if let Expr::Literal(val) = elem_expr {
-                                result_list.push(val);
-                            } else {
-                                // Return the literal list for now
-                                return Ok(Expr::Literal(crate::value::list(result_list)));
-                            }
-                        }
-                        Ok(Expr::Literal(crate::value::list(result_list)))
-                    }
-                }
-            } else {
-                // Non-symbol head - just quote the whole thing
-                Ok(Expr::Literal(value.clone()))
-            }
-        }
-
-        Value::Vector(_) => Ok(Expr::Literal(value.clone())),
-
-        // Cannot quote these
-        Value::Closure(_)
-        | Value::NativeFn(_)
-        | Value::LibHandle(_)
-        | Value::CHandle(_)
-        | Value::Exception(_)
-        | Value::Condition(_)
-        | Value::ThreadHandle(_) => Err("Cannot quote closure or native function".to_string()),
-    }
-}
-
-/// Simple value-to-expr conversion for bootstrap
-/// This is a simple tree-walking approach before full macro expansion
-pub fn value_to_expr(value: &Value, symbols: &mut SymbolTable) -> Result<Expr, String> {
-    let mut expr = value_to_expr_with_scope(value, symbols, &mut Vec::new())?;
-    super::capture_resolution::resolve_captures(&mut expr);
-    mark_tail_calls(&mut expr, true);
-    Ok(expr)
-}
-
 /// Convert a value to an expression, tracking local variable scopes
 /// The scope_stack contains local bindings (as Vec for ordering) at each nesting level
-fn value_to_expr_with_scope(
+pub fn value_to_expr_with_scope(
     value: &Value,
     symbols: &mut SymbolTable,
     scope_stack: &mut Vec<Vec<SymbolId>>,
@@ -544,13 +224,10 @@ fn value_to_expr_with_scope(
                     }
 
                     "let" => {
-                        // Syntax: (let ((var1 expr1) (var2 expr2) ...) body...)
-                        // Transform to: ((lambda (var1 var2 ...) body...) expr1 expr2 ...)
                         if list.len() < 2 {
                             return Err("let requires at least a binding vector".to_string());
                         }
 
-                        // Parse the bindings vector
                         let bindings_vec = list[1].list_to_vec()?;
                         let mut param_syms = Vec::new();
                         let mut binding_exprs = Vec::new();
@@ -565,15 +242,11 @@ fn value_to_expr_with_scope(
                             let var = binding_list[0].as_symbol()?;
                             param_syms.push(var);
 
-                            // Parse the binding expression in the current scope
-                            // (bindings cannot reference previous bindings or let-bound variables)
                             let expr =
                                 value_to_expr_with_scope(&binding_list[1], symbols, scope_stack)?;
                             binding_exprs.push(expr);
                         }
 
-                        // Parse the body (one or more expressions)
-                        // Body can reference let-bound variables, so we add them to scope
                         scope_stack.push(param_syms.clone());
 
                         let body_exprs: Result<Vec<_>, _> = list[2..]
@@ -592,33 +265,28 @@ fn value_to_expr_with_scope(
                             Box::new(Expr::Begin(body_exprs))
                         };
 
-                        // Analyze free variables in the body
                         let mut local_bindings = std::collections::HashSet::new();
                         for param in &param_syms {
                             local_bindings.insert(*param);
                         }
                         let free_vars = analyze_free_vars(&body, &local_bindings);
 
-                        // Convert free vars to captures, resolving their scope location
                         let mut sorted_free_vars: Vec<_> = free_vars.iter().copied().collect();
-                        sorted_free_vars.sort(); // Deterministic ordering
+                        sorted_free_vars.sort();
 
                         let captures: Vec<_> = sorted_free_vars
                             .iter()
                             .map(|sym| {
-                                // Look up in scope stack to determine if global or local
                                 for (reverse_idx, scope) in scope_stack.iter().enumerate().rev() {
                                     if let Some(local_index) = scope.iter().position(|s| s == sym) {
                                         let depth = scope_stack.len() - 1 - reverse_idx;
                                         return (*sym, depth, local_index);
                                     }
                                 }
-                                // If not found in scope stack, it's a global variable
                                 (*sym, 0, usize::MAX)
                             })
                             .collect();
 
-                        // Dead capture elimination: filter out captures that aren't actually used in the body
                         let candidates: HashSet<SymbolId> =
                             captures.iter().map(|(sym, _, _)| *sym).collect();
                         let actually_used =
@@ -628,18 +296,15 @@ fn value_to_expr_with_scope(
                             .filter(|(sym, _, _)| actually_used.contains(sym))
                             .collect();
 
-                        // Adjust variable indices in body to account for closure environment layout
                         let mut adjusted_body = body;
                         adjust_var_indices(&mut adjusted_body, &captures, &param_syms);
 
-                        // Create lambda: (lambda (var1 var2 ...) body...)
                         let lambda = Expr::Lambda {
                             params: param_syms,
                             body: adjusted_body,
                             captures,
                         };
 
-                        // Create call: (lambda expr1 expr2 ...)
                         Ok(Expr::Call {
                             func: Box::new(lambda),
                             args: binding_exprs,
@@ -648,13 +313,6 @@ fn value_to_expr_with_scope(
                     }
 
                     "let*" => {
-                        // Syntax: (let* ((var1 expr1) (var2 expr2) ...) body...)
-                        // Let* differs from let in that each binding can reference previous bindings.
-                        //
-                        // Strategy: parse binding expressions sequentially, adding each variable
-                        // to scope as we go. This naturally handles the sequential evaluation.
-                        // Then create a single large lambda with all parameters, using the
-                        // parsed expressions as arguments.
                         if list.len() < 2 {
                             return Err("let* requires at least a binding vector".to_string());
                         }
@@ -662,7 +320,6 @@ fn value_to_expr_with_scope(
                         let bindings_vec = list[1].list_to_vec()?;
 
                         if bindings_vec.is_empty() {
-                            // (let* () body...) - just evaluate body
                             let body_exprs: Result<Vec<_>, _> = list[2..]
                                 .iter()
                                 .map(|v| value_to_expr_with_scope(v, symbols, scope_stack))
@@ -677,7 +334,6 @@ fn value_to_expr_with_scope(
                             }
                         }
 
-                        // Parse bindings sequentially with growing scope
                         let mut param_syms = Vec::new();
                         let mut binding_exprs = Vec::new();
                         scope_stack.push(Vec::new());
@@ -692,19 +348,15 @@ fn value_to_expr_with_scope(
                             let var = binding_list[0].as_symbol()?;
                             param_syms.push(var);
 
-                            // Parse binding expression WITH PREVIOUS BINDINGS IN SCOPE
-                            // This allows y = (+ x 1) where x was previously bound
                             let expr =
                                 value_to_expr_with_scope(&binding_list[1], symbols, scope_stack)?;
                             binding_exprs.push(expr);
 
-                            // Add this variable to scope for next binding
                             if let Some(current_scope) = scope_stack.last_mut() {
                                 current_scope.push(var);
                             }
                         }
 
-                        // Parse body with all let* variables in scope
                         let body_exprs: Result<Vec<_>, _> = list[2..]
                             .iter()
                             .map(|v| value_to_expr_with_scope(v, symbols, scope_stack))
@@ -721,33 +373,28 @@ fn value_to_expr_with_scope(
                             Box::new(Expr::Begin(body_exprs))
                         };
 
-                        // Analyze free variables
                         let mut local_bindings = std::collections::HashSet::new();
                         for param in &param_syms {
                             local_bindings.insert(*param);
                         }
                         let free_vars = analyze_free_vars(&body, &local_bindings);
 
-                        // Convert free vars to captures, resolving their scope location
                         let mut sorted_free_vars: Vec<_> = free_vars.iter().copied().collect();
-                        sorted_free_vars.sort(); // Deterministic ordering
+                        sorted_free_vars.sort();
 
                         let captures: Vec<_> = sorted_free_vars
                             .iter()
                             .map(|sym| {
-                                // Look up in scope stack to determine if global or local
                                 for (reverse_idx, scope) in scope_stack.iter().enumerate().rev() {
                                     if let Some(local_index) = scope.iter().position(|s| s == sym) {
                                         let depth = scope_stack.len() - 1 - reverse_idx;
                                         return (*sym, depth, local_index);
                                     }
                                 }
-                                // If not found in scope stack, it's a global variable
                                 (*sym, 0, usize::MAX)
                             })
                             .collect();
 
-                        // Dead capture elimination: filter out captures that aren't actually used in the body
                         let candidates: HashSet<SymbolId> =
                             captures.iter().map(|(sym, _, _)| *sym).collect();
                         let actually_used =
@@ -757,18 +404,15 @@ fn value_to_expr_with_scope(
                             .filter(|(sym, _, _)| actually_used.contains(sym))
                             .collect();
 
-                        // Adjust variable indices in body to account for closure environment layout
                         let mut adjusted_body = body;
                         adjust_var_indices(&mut adjusted_body, &captures, &param_syms);
 
-                        // Create lambda: (lambda (var1 var2 ...) body...)
                         let lambda = Expr::Lambda {
                             params: param_syms,
                             body: adjusted_body,
                             captures,
                         };
 
-                        // Create call: (lambda expr1 expr2 ...)
                         Ok(Expr::Call {
                             func: Box::new(lambda),
                             args: binding_exprs,
@@ -777,9 +421,6 @@ fn value_to_expr_with_scope(
                     }
 
                     "letrec" => {
-                        // Syntax: (letrec ((var1 expr1) (var2 expr2) ...) body...)
-                        // All bindings are visible to all binding expressions and the body.
-                        // Unlike let, bindings can reference each other.
                         if list.len() < 2 {
                             return Err("letrec requires at least a binding vector".to_string());
                         }
@@ -788,7 +429,6 @@ fn value_to_expr_with_scope(
                         let mut param_syms = Vec::new();
                         let mut binding_exprs = Vec::new();
 
-                        // First pass: collect all variable names
                         for binding in &bindings_vec {
                             let binding_list = binding.list_to_vec()?;
                             if binding_list.len() != 2 {
@@ -799,9 +439,6 @@ fn value_to_expr_with_scope(
                             param_syms.push(binding_list[0].as_symbol()?);
                         }
 
-                        // Second pass: parse binding expressions
-                        // Names are NOT in scope during binding expression parsing
-                        // They'll reference each other as GlobalVar, resolved at runtime via scope_stack
                         for binding in &bindings_vec {
                             let binding_list = binding.list_to_vec()?;
                             let expr =
@@ -809,7 +446,6 @@ fn value_to_expr_with_scope(
                             binding_exprs.push(expr);
                         }
 
-                        // Parse body in current scope (names also as GlobalVar)
                         let body_exprs: Result<Vec<_>, _> = list[2..]
                             .iter()
                             .map(|v| value_to_expr_with_scope(v, symbols, scope_stack))
@@ -837,9 +473,8 @@ fn value_to_expr_with_scope(
                         let value =
                             Box::new(value_to_expr_with_scope(&list[2], symbols, scope_stack)?);
 
-                        // Look up the variable in the scope stack to determine depth and index
                         let mut depth = 0;
-                        let mut index = usize::MAX; // Use MAX to signal global variable
+                        let mut index = usize::MAX;
 
                         for (reverse_idx, scope) in scope_stack.iter().enumerate().rev() {
                             if let Some(local_index) = scope.iter().position(|sym| sym == &var) {
@@ -848,8 +483,6 @@ fn value_to_expr_with_scope(
                                 break;
                             }
                         }
-
-                        // If not found in local scopes (index == usize::MAX), it's a global variable set
 
                         Ok(Expr::Set {
                             var,
@@ -860,7 +493,6 @@ fn value_to_expr_with_scope(
                     }
 
                     "try" => {
-                        // Syntax: (try <body> (catch <var> <handler>) (finally <expr>)?)
                         if list.len() < 2 {
                             return Err("try requires at least a body".to_string());
                         }
@@ -870,7 +502,6 @@ fn value_to_expr_with_scope(
                         let mut catch_clause = None;
                         let mut finally_clause = None;
 
-                        // Parse catch and finally clauses
                         for item in &list[2..] {
                             if item.is_list() {
                                 let v = item.list_to_vec()?;
@@ -927,8 +558,6 @@ fn value_to_expr_with_scope(
                     }
 
                     "handler-case" => {
-                        // Syntax: (handler-case body (exception-id (var) handler-code) ...)
-                        // exception-id can be numeric or a symbol (will be looked up)
                         if list.len() < 2 {
                             return Err("handler-case requires at least a body".to_string());
                         }
@@ -937,7 +566,6 @@ fn value_to_expr_with_scope(
                             Box::new(value_to_expr_with_scope(&list[1], symbols, scope_stack)?);
                         let mut handlers = Vec::new();
 
-                        // Parse handler clauses
                         for clause in &list[2..] {
                             if let Ok(clause_vec) = clause.list_to_vec() {
                                 if clause_vec.len() != 3 {
@@ -947,11 +575,9 @@ fn value_to_expr_with_scope(
                                     );
                                 }
 
-                                // Parse exception ID
                                 let exception_id = match &clause_vec[0] {
                                     Value::Int(id) => *id as u32,
                                     Value::Symbol(sym) => {
-                                        // Map symbol to exception ID
                                         let name = symbols.name(*sym).unwrap_or("unknown");
                                         match name {
                                             "condition" => 1,
@@ -977,10 +603,8 @@ fn value_to_expr_with_scope(
                                     }
                                 };
 
-                                // Parse variable
                                 let var = clause_vec[1].as_symbol()?;
 
-                                // Parse handler code
                                 let handler_code = Box::new(value_to_expr_with_scope(
                                     &clause_vec[2],
                                     symbols,
@@ -997,8 +621,6 @@ fn value_to_expr_with_scope(
                     }
 
                     "handler-bind" => {
-                        // Syntax: (handler-bind ((exception-id handler-fn) ...) body)
-                        // handler-fn is called but doesn't unwind the stack
                         if list.len() != 3 {
                             return Err("handler-bind requires ((handlers...) body)".to_string());
                         }
@@ -1015,7 +637,6 @@ fn value_to_expr_with_scope(
                                 );
                             }
 
-                            // Parse exception ID
                             let exception_id = match &spec_vec[0] {
                                 Value::Int(id) => *id as u32,
                                 Value::Symbol(sym) => {
@@ -1039,7 +660,6 @@ fn value_to_expr_with_scope(
                                 }
                             };
 
-                            // Parse handler function
                             let handler_fn = Box::new(value_to_expr_with_scope(
                                 &spec_vec[1],
                                 symbols,
@@ -1056,7 +676,6 @@ fn value_to_expr_with_scope(
                     }
 
                     "match" => {
-                        // Syntax: (match value (pattern1 result1) (pattern2 result2) ... [default])
                         if list.len() < 2 {
                             return Err("match requires at least a value".to_string());
                         }
@@ -1066,51 +685,42 @@ fn value_to_expr_with_scope(
                         let mut patterns = Vec::new();
                         let mut default = None;
 
-                        // Parse pattern clauses
                         for clause in &list[2..] {
                             if let Ok(clause_vec) = clause.list_to_vec() {
                                 if clause_vec.is_empty() {
                                     return Err("Empty pattern clause".to_string());
                                 }
 
-                                // Check if this is a default clause (symbol, not a list)
                                 if clause_vec.len() == 1 {
-                                    // Single value - treat as default
                                     default = Some(Box::new(value_to_expr_with_scope(
                                         &clause_vec[0],
                                         symbols,
                                         scope_stack,
                                     )?));
                                 } else if clause_vec.len() == 2 {
-                                    // Pattern and result
-                                    let pattern = value_to_pattern(&clause_vec[0], symbols)?;
+                                    let pattern = super::super::patterns::value_to_pattern(
+                                        &clause_vec[0],
+                                        symbols,
+                                    )?;
 
-                                    // Extract pattern variables
                                     let pattern_vars = extract_pattern_variables(&pattern);
 
-                                    // If there are pattern variables, wrap the body in a lambda
-                                    // that binds them to the matched value
                                     let result = if !pattern_vars.is_empty() {
-                                        // Add pattern variables to scope for parsing the body
                                         let mut new_scope_stack = scope_stack.clone();
                                         new_scope_stack.push(pattern_vars.clone());
 
-                                        // Parse the result in the new scope
                                         let body_expr = value_to_expr_with_scope(
                                             &clause_vec[1],
                                             symbols,
                                             &mut new_scope_stack,
                                         )?;
 
-                                        // Transform: (lambda (var1 var2 ...) body_expr)
-                                        // This binds the pattern variables for use in the body
                                         Expr::Lambda {
                                             params: pattern_vars,
                                             body: Box::new(body_expr),
                                             captures: Vec::new(),
                                         }
                                     } else {
-                                        // No variables to bind
                                         value_to_expr_with_scope(
                                             &clause_vec[1],
                                             symbols,
@@ -1137,13 +747,9 @@ fn value_to_expr_with_scope(
                     }
 
                     "throw" => {
-                        // Syntax: (throw <exception>)
-                        // Throw is a special form that compiles to a function call
-                        // The throw primitive will convert the exception to a Rust error
                         if list.len() != 2 {
                             return Err("throw requires exactly 1 argument".to_string());
                         }
-                        // Compile as a regular function call to the throw primitive
                         let func = Box::new(Expr::GlobalVar(first.as_symbol()?));
                         let args = vec![value_to_expr_with_scope(&list[1], symbols, scope_stack)?];
                         Ok(Expr::Call {
@@ -1154,8 +760,6 @@ fn value_to_expr_with_scope(
                     }
 
                     "defmacro" | "define-macro" => {
-                        // Syntax: (defmacro name (params...) body)
-                        //      or (define-macro name (params...) body)
                         if list.len() != 4 {
                             return Err(
                                 "defmacro requires exactly 3 arguments (name, parameters, body)"
@@ -1165,7 +769,6 @@ fn value_to_expr_with_scope(
                         let name = list[1].as_symbol()?;
                         let params_val = &list[2];
 
-                        // Parse parameter list
                         let params = if params_val.is_list() {
                             let param_vec = params_val.list_to_vec()?;
                             param_vec
@@ -1176,10 +779,8 @@ fn value_to_expr_with_scope(
                             return Err("Macro parameters must be a list".to_string());
                         };
 
-                        // Store macro body as source code for later expansion
                         let body_str = format!("{}", list[3]);
 
-                        // Register the macro in the symbol table
                         use crate::symbol::MacroDef;
                         symbols.define_macro(MacroDef {
                             name,
@@ -1194,7 +795,6 @@ fn value_to_expr_with_scope(
                     }
 
                     "while" => {
-                        // Syntax: (while condition body)
                         if list.len() != 3 {
                             return Err(
                                 "while requires exactly 2 arguments (condition body)".to_string()
@@ -1208,8 +808,6 @@ fn value_to_expr_with_scope(
                     }
 
                     "for" => {
-                        // Syntax: (for var iter body)
-                        // Also supports: (for var in iter body) for clarity
                         if list.len() < 4 || list.len() > 5 {
                             return Err(
                                 "for requires 3 or 4 arguments (var [in] iter body)".to_string()
@@ -1218,27 +816,23 @@ fn value_to_expr_with_scope(
 
                         let var = list[1].as_symbol()?;
                         let (iter_expr, body_expr) = if list.len() == 4 {
-                            // (for var iter body)
                             (&list[2], &list[3])
-                        } else {
-                            // (for var in iter body)
-                            if let Value::Symbol(in_sym) = &list[2] {
-                                if let Some("in") = symbols.name(*in_sym) {
-                                    (&list[3], &list[4])
-                                } else {
-                                    return Err("for loop syntax: (for var iter body) or (for var in iter body)".to_string());
-                                }
+                        } else if let Value::Symbol(in_sym) = &list[2] {
+                            if let Some("in") = symbols.name(*in_sym) {
+                                (&list[3], &list[4])
                             } else {
                                 return Err("for loop syntax: (for var iter body) or (for var in iter body)".to_string());
                             }
+                        } else {
+                            return Err(
+                                "for loop syntax: (for var iter body) or (for var in iter body)"
+                                    .to_string(),
+                            );
                         };
 
-                        // Compile iterator expression
                         let iter =
                             Box::new(value_to_expr_with_scope(iter_expr, symbols, scope_stack)?);
 
-                        // Compile body expression (the loop variable will be set as a global at runtime)
-                        // Note: the loop variable is accessible in the body as a global
                         let body =
                             Box::new(value_to_expr_with_scope(body_expr, symbols, scope_stack)?);
 
@@ -1246,10 +840,8 @@ fn value_to_expr_with_scope(
                     }
 
                     "and" => {
-                        // Syntax: (and expr1 expr2 ...)
-                        // Short-circuit evaluation: returns first falsy value or last value
                         if list.len() < 2 {
-                            return Ok(Expr::Literal(Value::Bool(true))); // (and) => true
+                            return Ok(Expr::Literal(Value::Bool(true)));
                         }
                         let exprs: Result<Vec<_>, _> = list[1..]
                             .iter()
@@ -1259,10 +851,8 @@ fn value_to_expr_with_scope(
                     }
 
                     "or" => {
-                        // Syntax: (or expr1 expr2 ...)
-                        // Short-circuit evaluation: returns first truthy value or last value
                         if list.len() < 2 {
-                            return Ok(Expr::Literal(Value::Bool(false))); // (or) => false
+                            return Ok(Expr::Literal(Value::Bool(false)));
                         }
                         let exprs: Result<Vec<_>, _> = list[1..]
                             .iter()
@@ -1272,11 +862,8 @@ fn value_to_expr_with_scope(
                     }
 
                     "xor" => {
-                        // Syntax: (xor expr1 expr2 ...)
-                        // Transform to a function call to the xor primitive
-                        // This way we don't need special compilation logic
                         if list.len() < 2 {
-                            return Ok(Expr::Literal(Value::Bool(false))); // (xor) => false
+                            return Ok(Expr::Literal(Value::Bool(false)));
                         }
 
                         let func = Box::new(Expr::GlobalVar(symbols.intern("xor")));
@@ -1292,23 +879,17 @@ fn value_to_expr_with_scope(
                     }
 
                     "->" => {
-                        // Syntax: (-> value form1 form2 ...)
-                        // Thread first: inserts value as the FIRST argument to each form
-                        // Example: (-> 5 (+ 10) (* 2)) => (* (+ 5 10) 2)
                         if list.len() < 2 {
                             return Err("-> requires at least a value and one form".to_string());
                         }
 
-                        // Start with the value
                         let mut result = value_to_expr_with_scope(&list[1], symbols, scope_stack)?;
 
-                        // Thread through each form
                         for form in &list[2..] {
                             result = if let Ok(form_vec) = form.list_to_vec() {
                                 if form_vec.is_empty() {
                                     return Err("Cannot thread through empty form".to_string());
                                 }
-                                // Insert result as the first argument
                                 let func =
                                     value_to_expr_with_scope(&form_vec[0], symbols, scope_stack)?;
                                 let mut args = vec![result];
@@ -1321,7 +902,6 @@ fn value_to_expr_with_scope(
                                     tail: false,
                                 }
                             } else {
-                                // If form is not a list, treat it as a function call with result as arg
                                 Expr::Call {
                                     func: Box::new(value_to_expr_with_scope(
                                         form,
@@ -1338,29 +918,23 @@ fn value_to_expr_with_scope(
                     }
 
                     "->>" => {
-                        // Syntax: (->> value form1 form2 ...)
-                        // Thread last: inserts value as the LAST argument to each form
-                        // Example: (->> 5 (+ 10) (* 2)) => (* 2 (+ 10 5))
                         if list.len() < 2 {
                             return Err("->> requires at least a value and one form".to_string());
                         }
 
-                        // Start with the value
                         let mut result = value_to_expr_with_scope(&list[1], symbols, scope_stack)?;
 
-                        // Thread through each form
                         for form in &list[2..] {
                             result = if let Ok(form_vec) = form.list_to_vec() {
                                 if form_vec.is_empty() {
                                     return Err("Cannot thread through empty form".to_string());
                                 }
-                                // Collect arguments up to the end, then add result as last
                                 let func =
                                     value_to_expr_with_scope(&form_vec[0], symbols, scope_stack)?;
-                                let mut args = Vec::new();
-                                for arg in &form_vec[1..] {
-                                    args.push(value_to_expr_with_scope(arg, symbols, scope_stack)?);
-                                }
+                                let mut args: Vec<_> = form_vec[1..]
+                                    .iter()
+                                    .map(|v| value_to_expr_with_scope(v, symbols, scope_stack))
+                                    .collect::<Result<_, _>>()?;
                                 args.push(result);
                                 Expr::Call {
                                     func: Box::new(func),
@@ -1368,7 +942,6 @@ fn value_to_expr_with_scope(
                                     tail: false,
                                 }
                             } else {
-                                // If form is not a list, treat it as a function call with result as arg
                                 Expr::Call {
                                     func: Box::new(value_to_expr_with_scope(
                                         form,
@@ -1385,11 +958,6 @@ fn value_to_expr_with_scope(
                     }
 
                     "cond" => {
-                        // Syntax: (cond (test1 body1) (test2 body2) ... [(else body)])
-                        // A cond expression evaluates test expressions in order until one is truthy,
-                        // then evaluates and returns its corresponding body.
-                        // If no tests are truthy and there's an else clause, evaluate the else body.
-                        // If no tests are truthy and there's no else clause, return nil.
                         if list.len() < 2 {
                             return Err("cond requires at least one clause".to_string());
                         }
@@ -1470,15 +1038,11 @@ fn value_to_expr_with_scope(
                         if let Value::Symbol(sym_id) = first {
                             if symbols.is_macro(*sym_id) {
                                 if let Some(macro_def) = symbols.get_macro(*sym_id) {
-                                    // This is a macro call - expand it
-                                    // Get the arguments as unevaluated values
                                     let args = list[1..].to_vec();
 
-                                    // Expand the macro
                                     let expanded =
                                         expand_macro(*sym_id, &macro_def, &args, symbols)?;
 
-                                    // Parse the expanded result as an expression
                                     return value_to_expr_with_scope(
                                         &expanded,
                                         symbols,
@@ -1517,119 +1081,5 @@ fn value_to_expr_with_scope(
         }
 
         _ => Err(format!("Cannot convert {:?} to expression", value)),
-    }
-}
-
-/// Mark Call expressions in tail position with tail=true for TCO
-fn mark_tail_calls(expr: &mut Expr, in_tail: bool) {
-    match expr {
-        Expr::Call { func, args, tail } => {
-            if in_tail {
-                *tail = true;
-            }
-            // Function and arguments are NOT in tail position
-            mark_tail_calls(func, false);
-            for arg in args {
-                mark_tail_calls(arg, false);
-            }
-        }
-        Expr::If { cond, then, else_ } => {
-            mark_tail_calls(cond, false);
-            mark_tail_calls(then, in_tail);
-            mark_tail_calls(else_, in_tail);
-        }
-        Expr::Cond { clauses, else_body } => {
-            for (test, body) in clauses {
-                mark_tail_calls(test, false);
-                mark_tail_calls(body, in_tail);
-            }
-            if let Some(else_expr) = else_body {
-                mark_tail_calls(else_expr, in_tail);
-            }
-        }
-        Expr::Begin(exprs) => {
-            let len = exprs.len();
-            for (i, e) in exprs.iter_mut().enumerate() {
-                let is_last = i == len - 1;
-                mark_tail_calls(e, in_tail && is_last);
-            }
-        }
-        Expr::Block(exprs) => {
-            let len = exprs.len();
-            for (i, e) in exprs.iter_mut().enumerate() {
-                let is_last = i == len - 1;
-                mark_tail_calls(e, in_tail && is_last);
-            }
-        }
-        Expr::Lambda { body, .. } => {
-            // Lambda body is in tail position (of the lambda)
-            mark_tail_calls(body, true);
-        }
-        Expr::Let { bindings, body } => {
-            for (_, e) in bindings {
-                mark_tail_calls(e, false);
-            }
-            mark_tail_calls(body, in_tail);
-        }
-        Expr::Letrec { bindings, body } => {
-            for (_, e) in bindings {
-                mark_tail_calls(e, false);
-            }
-            mark_tail_calls(body, in_tail);
-        }
-        Expr::Set { value, .. } => {
-            mark_tail_calls(value, false);
-        }
-        Expr::Define { value, .. } => {
-            mark_tail_calls(value, false);
-        }
-        Expr::While { cond, body } => {
-            mark_tail_calls(cond, false);
-            mark_tail_calls(body, false); // Loop body is not in tail position
-        }
-        Expr::For { iter, body, .. } => {
-            mark_tail_calls(iter, false);
-            mark_tail_calls(body, false);
-        }
-        Expr::Match {
-            value,
-            patterns,
-            default,
-        } => {
-            mark_tail_calls(value, false);
-            for (_, body) in patterns {
-                mark_tail_calls(body, in_tail);
-            }
-            if let Some(d) = default {
-                mark_tail_calls(d, in_tail);
-            }
-        }
-        Expr::Try {
-            body,
-            catch,
-            finally,
-        } => {
-            // Try body is NOT in tail position (might need to run finally)
-            mark_tail_calls(body, false);
-            if let Some((_, handler)) = catch {
-                mark_tail_calls(handler, false);
-            }
-            if let Some(f) = finally {
-                mark_tail_calls(f, false);
-            }
-        }
-        Expr::And(exprs) | Expr::Or(exprs) => {
-            // Only the last expression in and/or is in tail position
-            let len = exprs.len();
-            for (i, e) in exprs.iter_mut().enumerate() {
-                let is_last = i == len - 1;
-                mark_tail_calls(e, in_tail && is_last);
-            }
-        }
-        Expr::DefMacro { body, .. } => {
-            mark_tail_calls(body, false);
-        }
-        // Leaf nodes and others — nothing to do
-        _ => {}
     }
 }
